extract LoginUseCaseOutputSpy to it's own file
